+++ src/renderer/components/SubtitleList.jsx
@@ -43,7 +43,7 @@ export default ({
     const [rememberAddCollection, setRememberAddCollection] = useState(false);
 
     let videoLengthMs = videoLength * 1000;
-    let defaultClipSize = videoLengthMs * 0.1;
+    let defaultClipSize = 8000; // 8 seconds
 
     useEffect(() => {
         getCollections();